chore(fhir): TEFCA QHIN Phase G — D-3 validator/types relocation + B-3 history triggers
Two follow-ups from earlier phases bundled into one PR. D-2 stays
deferred until write paths land.

D-3: validator + FHIR type definitions move to supabase/functions/_shared/

Before: tefca-ias/validation.ts imported the validator from
../../../src/services/fhir/validation/core.ts, leaving the runtime path
conditional on Supabase Edge bundling following arbitrary upward imports.

After:
  supabase/functions/_shared/fhir/types.ts                 (canonical)
  supabase/functions/_shared/fhir-validation/core.ts       (canonical)

  src/services/fhir/types.ts            -> 1-line re-export shim
  src/services/fhir/validation/core.ts  -> 1-line re-export shim
  src/services/fhir/uscore-validator.ts -> existing re-export still works,
                                            now via the new chain
  tefca-ias/validation.ts               -> imports ../_shared/fhir-validation/core.ts

Net: the Deno-side import is canonical (no cross src <-> supabase
boundary at the function root); the browser side crosses the boundary
in the cheaper direction (src -> _shared) which Vite + tsc both resolve
cleanly. Existing browser callers (resourceMapper.ts, dexieAdapters.ts,
dexieExporter.ts, bundleGenerator.ts, tefcaAuth.ts, db migration-runner)
keep importing from "./types" and need no changes.

B-3: extend resource_versions snapshot triggers to remaining clinical
resources

Phase B-2 covered the six clinician-mutated tables that have a clean
1-row -> 1-FHIR-resource mapping (patients, conditions, prescriptions,
patient_allergies, care_plans, goals).

This migration adds AFTER INSERT/UPDATE/DELETE triggers wired into the
existing fhir_record_resource_version() function for the six remaining
single-row -> single-FHIR resources:

  visits               -> Encounter
  immunizations        -> Immunization
  procedures           -> Procedure
  document_references  -> DocumentReference
  service_requests     -> ServiceRequest
  dispenses            -> MedicationDispense (+ MedicationRequest reuses
                                                the same snapshot via
                                                history.ts:loadResourceVersion's
                                                mapper indirection)

Registry update: bumps interactions for those six resources from
baseInteractions (read, search-type) to versionedInteractions
(read, vread, search-type, history-instance). The CapabilityStatement
generator picks the change up automatically. /metadata, /Resource/{id}/_history,
and /Resource/{id}/_history/{vid} all work for these resource types
once this migration is applied.

Intentionally not covered: Observation (vitals + sdoh_observations) and
DiagnosticReport (lab_orders + lab_results join). Both have row-to-resource
expansion that a single-row snapshot can't faithfully represent. The
migration documents the caveat.

Verification
- npx tsc --noEmit clean (cross-boundary re-exports resolve)
- npx eslint --max-warnings=0 clean
- src/services tests pass (29 tests)

Open follow-ups
- Phase D-2: validator-gated write paths once write endpoints land.
- Per-column snapshot strategy for Observation + DiagnosticReport so
  those two also gain _history coverage.
